refactor: split controllers part 2 and base url change to drgapi
