+++ Controllers/Top10GamesController.cs
@@ -6,7 +6,7 @@ using DRG_Api.Services;
 
 namespace DRG_Api.Controllers
 {
-    [Route("api/PlayedGames")]
+    [Route("drgapi/[controller]")]
     [ApiController]
     public class Top10GamesController : ControllerBase
     {
@@ -17,13 +17,13 @@ namespace DRG_Api.Controllers
             _repositories = unitOfWork;
         }
 
-        [HttpGet("top10games")]
+        [HttpGet]
         public async Task<ActionResult<IEnumerable<Top10Game>>> GetTop10Games()
         {
             return await _repositories.Top10Games.FindAll();
         }
         
-        [HttpGet("top10games/{name}")]
+        [HttpGet("{name}")]
         public async Task<ActionResult<IEnumerable<Top10Game>>> GetTop10GamesBy(string name)
         {
             var top10name = await _repositories.Top10Names.FindOneBy(top10name => top10name.name.Contains(name));
@@ -32,7 +32,7 @@ namespace DRG_Api.Controllers
             return top10;
         }
 
-        [HttpPost("top10games/{top10name}")]
+        [HttpPost("{top10name}")]
         public async Task<ActionResult<Top10Game>> PostTop10Game(Top10Game game, string top10name)
         {
             var top10nameObject = await _repositories.Top10Names.FindOneBy(top10 => top10.name.Contains(top10name));
@@ -42,7 +42,7 @@ namespace DRG_Api.Controllers
             return CreatedAtAction(nameof(GetTop10Games), new { id = game.id }, game);
         }
 
-        [HttpPut("top10games/{id}")]
+        [HttpPut("{id}")]
         public async Task<IActionResult> PutTop10Game(int id, Top10Game game)
         {
             if (id != game.id)
@@ -63,7 +63,7 @@ namespace DRG_Api.Controllers
             return NoContent();
         }
 
-        [HttpDelete("top10games/{id}")]
+        [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteTop10Game(int id)
         {
             var game = await _repositories.Top10Games.Find(id);
@@ -73,52 +73,6 @@ namespace DRG_Api.Controllers
             }
 
             await _repositories.Top10Games.DeleteAsync(game);
-
-            return NoContent();
-        }
-
-        [HttpGet("top10names")]
-        public async Task<ActionResult<IEnumerable<Top10Name>>> GetTop10Names()
-        {
-            return await _repositories.Top10Names.FindAll();
-        }
-        [HttpPost("top10names")]
-        public async Task<ActionResult<Top10Name>> PostTop10Name(Top10Name top10)
-        {
-            await _repositories.Top10Names.CreateAsync(top10);
-
-            return CreatedAtAction(nameof(GetTop10Names), new { id = top10.id }, top10);
-        }
-        [HttpPut("top10names/{id}")]
-        public async Task<IActionResult> PutTop10Name(int id, Top10Name name)
-        {
-            if (id != name.id)
-            {
-                return BadRequest();
-            }
-
-            try
-            {
-                await _repositories.Top10Names.UpdateAsync(name);
-            }
-            catch
-            {
-                return NotFound();
-            }
-
-
-            return NoContent();
-        }
-        [HttpDelete("top10names/{id}")]
-        public async Task<IActionResult> DeleteTop10Name(int id)
-        {
-            var name = await _repositories.Top10Names.Find(id);
-            if (name == null)
-            {
-                return NotFound();
-            }
-
-            await _repositories.Top10Names.DeleteAsync(name);
 
             return NoContent();
         }
